resize

+++ include/candidate_window.hpp
@@ -32,7 +32,7 @@ class CandidateWindow {
     virtual void set_highlight_callback(std::function<void(size_t index)>) = 0;
     virtual void set_select_callback(std::function<void(size_t index)>) = 0;
     virtual void set_style(const void *style) = 0;
-    virtual void show(float x, float y) = 0;
+    virtual void show(double x, double y) = 0;
     virtual void hide() = 0;
 };
 } // namespace candidate_window
